backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
email-validator==2.2.0